@@ -4,11 +4,15 @@ namespace App\Form;
 
 use App\Entity\Lesson;
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\DateType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\Extension\Core\Type\UrlType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+
+use Symfony\Component\Validator\Constraints as Assert;
 
 class LessonType extends AbstractType
 {
@@ -16,14 +20,21 @@ class LessonType extends AbstractType
     {
         $builder
             ->add('title', TextType::class, [
-                'label' => 'Lesson title'
+                'label' => 'Lesson title',
+                'constraints' => [
+                    new Assert\Length(['min' => 2, 'max' => 75]),
+                    new Assert\NotBlank()
+                ]
             ])
-            ->add('video', TextType::class)
-            ->add('content', TextareaType::class)
-            ->add('publishedAt', DateType::class, [
-                'widget' => 'single_text'
+            ->add('video', UrlType::class)
+            ->add('content', TextareaType::class, [
+                'constraints' => [
+                    new Assert\NotNull()
+                ]
             ])
-            ->add('isPublished')
+            ->add('isPublished', CheckboxType::class, [
+                'required' => false
+            ])
         ;
     }
 
