@@ -3,14 +3,13 @@
 namespace App\Form;
 
 use App\Entity\Section;
-use App\entity\Lesson;
 use App\Form\LessonType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\CollectionType;
-use Symfony\Component\Form\Extension\Core\Type\DateType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Validator\Constraints as Assert;
 
 class SectionType extends AbstractType
 {
@@ -18,10 +17,11 @@ class SectionType extends AbstractType
     {
         $builder
             ->add('title', TextType::class, [
-                'label' => 'Section title'
-            ])
-            ->add('publishedAt', DateType::class, [
-                'widget' => 'single_text'
+                'label' => 'Section title',
+                'constraints' => [
+                    new Assert\NotBlank(),
+                    new Assert\Length(['min' => 2, 'max' => 75])
+                ]
             ])
             ->add('lessons', CollectionType::class, [
                 'label' => false,
